requirements.txt
Django>=3.2,<4.0
djangorestframework==3.12.4
psycopg2-binary>=2.8
django-simple-history>=2.10.0
django-rest-framework>=0.1.0
drf-yasg>=1.20.0
pylint-django>=2.4.4
django-import-export>=2.5.0